Allow ordering on filtered products.
This allows the list of products to be filtered by Artist and still be able to reorder both within that Artist and retain correct overall menu ordering. It does this by gathering the menu orders of all of the products that are shown for that Artist, and only reorders the products using those menu orders. Amazing!